requirements.txt
tqdm==4.66.5
joblib==1.4.2
Pyyaml==6.0.2
tifffile==2023.4.12
natsort==7.1.1
matplotlib==3.7.2
opencv-python==4.11.0.86
numpy==1.24.3
scikit-learn==1.3.0
scikit-image==0.20.0
torch==1.10.1
torchvision==0.11.2
Pillow==10.4.0
pandas==2.0.3
openpyxl==3.1.5